# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restau TyTy API",
        "description": "API documentation for Restaurant Management System",
        "contact": {
            "email": "admin@restautyty.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://restau-bo.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/tables/{tableId}/commandes": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Get list of commandes",
                "operationId": "getCommandesList",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Store new commande",
                "description": "Creates a new order. If guest, returns a session token.",
                "operationId": "storeCommande",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commande"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Commande"
                                        },
                                        "token": {
                                            "description": "Session token for guest (UUID)",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tables/{tableId}/commandes/{id}": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Get commande information",
                "description": "For guests, requires X-Order-Token header matching the order.",
                "operationId": "getCommandeById",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Order-Token",
                        "in": "header",
                        "description": "Guest Session Token (required if not authenticated)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Update commande",
                "operationId": "updateCommande",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commande"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Delete commande",
                "operationId": "deleteCommande",
                "parameters": [
                    {
                        "name": "tableId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Commande": {
                "required": [
                    "table_id",
                    "statut"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "table_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "statut": {
                        "type": "string",
                        "example": "en_attente"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}